refactor deps
