+++ events/blocks/events-form/events-form.js
@@ -1,8 +1,8 @@
 import { LIBS } from '../../scripts/scripts.js';
 import { getMetadata } from '../../scripts/utils.js';
-import HtmlSanitizer from '../../deps/html-sanitizer.js';
+import HtmlSanitizer from '../../scripts/deps/html-sanitizer.js';
 import { createAttendee, deleteAttendee, updateAttendee } from '../../scripts/esp-controller.js';
-import BlockMediator from '../../deps/block-mediator.min.js';
+import BlockMediator from '../../scripts/deps/block-mediator.min.js';
 
 const { createTag } = await import(`${LIBS}/utils/utils.js`);
 const { closeModal } = await import(`${LIBS}/blocks/modal/modal.js`);
